# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  gitlab_com_xamops_auth_internal_dto.Permission:
    properties:
      description:
        type: string
      title:
        type: string
    type: object
  gitlab_com_xamops_auth_internal_dto.Role:
    properties:
      description:
        type: string
      permissions:
        items:
          $ref: '#/definitions/gitlab_com_xamops_auth_internal_dto.Permission'
        type: array
      title:
        type: string
    type: object
  gitlab_com_xamops_auth_internal_dto.User:
    properties:
      email:
        type: string
      is_active:
        type: boolean
      organization_id:
        type: string
      role:
        $ref: '#/definitions/gitlab_com_xamops_auth_internal_dto.Role'
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Swagger xamops auth API
  version: 1.0.0
paths:
...
  /user/current:
    get:
      consumes:
      - application/json
      description: Get Current User
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gitlab_com_xamops_auth_internal_dto.User'
      security:
      - ApiKeyAuth: []
      summary: Get Current User
      tags:
      - User
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"